@@ -10,12 +10,20 @@
   },
   "dependencies": {
     "date-fns": "^3.6.0",
+    "icons": "^1.0.0",
     "moment": "^2.30.1",
     "mongoose": "^8.4.0",
-    "next": "14.2.3",
-    "react": "^18",
-    "react-dom": "^18",
+    "next": "^14.2.5",
+    "nodemailer": "^6.9.14",
+    "react": "^18.3.1",
+    "react-datepicker": "^7.0.0",
+    "react-dom": "^18.3.1",
     "react-icons": "^5.2.1",
-    "sass": "^1.77.2"
+    "sass": "^1.77.2",
+    "zustand": "^4.5.2"
+  },
+  "devDependencies": {
+    "eslint": "8.57.0",
+    "eslint-config-next": "14.2.3"
   }
 }
