@@ -18,7 +18,7 @@ const UserSchema = new mongoose.Schema({
     orders: [
         {
             type: mongoose.Schema.Types.ObjectId,
-            ref: 'orders'
+            ref: 'Orders'
         }
     ],
     type: {
@@ -29,5 +29,9 @@ const UserSchema = new mongoose.Schema({
 })
 
 
-const UserModel = mongoose.models.Users || mongoose.model('Users', UserSchema);
-module.exports = UserModel;
+// const UserModel = mongoose.models.Users || mongoose.model('Users', UserSchema);
+// module.exports = UserModel;
+
+export const UserModel = mongoose.models?.Users || mongoose.model('Users', UserSchema)
+
+// module.exports = mongoose.models?.Users || mongoose.model('Users', UserSchema);
