@@ -2,16 +2,29 @@ import { connectToMongo } from '@/server/connectToMongo'
 import { getShowsByArtist } from '@/server/shows/show.service'
 import styles from './style.module.scss'
 import Link from 'next/link'
+import { getArtistById } from '@/server/artists/artist.service'
 
 export default async function page({ params }) {
 
   await connectToMongo()
-  const artistName = params.artistName
-  const shows = await getShowsByArtist(artistName)
+  const artistId = params.artistId
+  const artist = await getArtistById(artistId)
+  console.log(artist);
+  const shows = await getShowsByArtist(artistId)
   // console.log(shows);
 
   return (
     <div className={styles.homeContainer}>
+      {/* <img height={420} src={artist.image} alt={artist.fullName} className={styles.img}/>  */}
+      <div className={styles.artistName}>{artist.fullName}</div>
+      <div className={styles.bodyContainer}>
+        <div className={styles.left}>
+          <div className={styles.about}>About</div>
+          <div>{artist.description}</div>
+        </div>
+        <div className={styles.right} ></div>
+      </div>
+      <div>{artist.shows.length}</div>
       <div className={styles.home}>
         {shows.map(show => {
           return <li key={show._id}>
