@@ -1,8 +1,13 @@
-import React from 'react'
+"use client"
+import React, { useState } from 'react'
 import styles from './style.module.scss'
-import GetTicket from '@/app/components/GetTicket'
-export default function index({ title, subTitel, price }) {
-   console.log({ title, subTitel, price });
+import { useStoreCount } from '@/utils/store'
+
+export default function PriceTicket({ title, subTitel, price }) {
+  
+   const count = useStoreCount(state=> state.count)
+   const increaseCount = useStoreCount(state=> state.increaseCount)
+   const decreaseCount = useStoreCount(state=> state.decreaseCount)
    return (
       <div className={styles.priceTicket}>
          <div className={styles.titels} >
@@ -11,8 +16,16 @@ export default function index({ title, subTitel, price }) {
          </div>
 
          <div className={styles.titelss} >
-         <span className={styles.price}>{price}</span>
-         <GetTicket/>
+            <span className={styles.price}>{price}$</span>
+            <div className={styles.getTicket}>
+               <button className={styles.btn} onClick={increaseCount} >
+                  +
+               </button>
+               <span className={styles.num}>{count}</span>
+               <button className={styles.btn} onClick={decreaseCount}>
+                  -
+               </button>
+            </div>
          </div>
 
       </div>
