@@ -1,31 +1,42 @@
-import React from 'react'
-import { getShowById } from '@/server/shows/show.service'
 import styles from './style.module.scss'
 import Link from 'next/link'
-import GetTicket from '@/app/components/GetTicket'
 import { IoChevronBackOutline, IoLocationSharp } from "react-icons/io5";
 import { IoMdClose } from "react-icons/io";
 import { FaRegCalendarCheck } from 'react-icons/fa'
-import moment from 'moment'
 import PriceTicket from '@/app/components/PriceTicket'
-import ContinueButton from '@/app/components/ContinueButton'
+import { connectToMongo } from '@/server/connectToMongo';
+import ContinueButton from '@/app/components/ContinueButton';
+import moment from 'moment';
+import { getFullShowById } from '@/server/shows/show.service';
 
-export default async function index({ params }) {
+export default async function page({ params }) {
+  await connectToMongo()
   const showId = params.showId
-  const show = await getShowById(showId)
-  const pic = show.picture_of_artist
-  const date = show.date
-  const formatedFullDate = moment(date).format('MMM D, YYYY h:mm A');
-  const titel = "Standart Ticket"///צריך לייבא מהדאטא
+  const show = await getFullShowById(showId)
+
+  const serializedShow = {
+    id: show._id.toString(),
+    artist: show.artist.fullName,
+    picture: show.picture_of_artist,
+    location: show.location,
+    date: show.date,
+    title: show.title,
+    price: show.price_per_ticket
+  };
+  
+
+  // const pic = show.picture_of_artist
+  // const date = show.date
+  // const price = show.price_per_ticket
+  const partDate = moment(serializedShow.date).format('MMM D, YYYY');
   const subTitel = "General admission entry into the event"///צריך לייבא מהדאטא
-  const price = show.price_per_ticket
 
   return (
     <div className={styles.purchContainer}>
       <div className={styles.bodyCont}>
 
         <div className={styles.pageHeader}>
-          <Link href={`/show/${show._id}`} className={styles.headerLeft}>
+          <Link href={`/show/${serializedShow.id}`} className={styles.headerLeft}>
             <IoChevronBackOutline className={styles.backIcon} />
             <span className={styles.headerTitle}>Select tickets</span>
           </Link>
@@ -36,34 +47,30 @@ export default async function index({ params }) {
 
         <div className={styles.pageBody}>
           <div className={styles.bodyLeft}>
-            <img src={pic} alt={show.artist} className={styles.picture} />
+            <img src={serializedShow.picture} alt={'show.artist'} className={styles.picture} />
             <div className={styles.details}>
-              {/* <span>{show.title}</span> */}
-              <span className={styles.artist} >{show.artist}</span>
+              <span className={styles.artist} >{serializedShow.artist}</span>
               <div className={styles.locationCont}>
                 <IoLocationSharp />
-                <span>{show.location}</span>
+                <span>{serializedShow.location}</span>
               </div>
 
               <div className={styles.dateCont}>
                 <FaRegCalendarCheck />
-                <span>{formatedFullDate}</span>
+                <span>{partDate}</span>
               </div>
             </div>
           </div>
           <div className={styles.bodyRight}>
             <h2>
-              {show.title}
+              {serializedShow.title}
             </h2>
             <div className={styles.priceTicket}>
-              <PriceTicket title={titel} subTitel={subTitel} price={price} />
+              <PriceTicket title={serializedShow.title} subTitel={subTitel} price={serializedShow.price} />
             </div>
-            <div>
-              <ContinueButton />
-            </div>
+            <ContinueButton showId={serializedShow.id} price={serializedShow.price}/>
           </div>
         </div>
-
       </div>
     </div>
   )
