@@ -1,28 +1,29 @@
 const mongoose = require('mongoose')
-
+import '../shows/show.model';
 
 const ArtistSchema = new mongoose.Schema({
-fullName: {
-    type: String,
-    required: true
-},
-image: {
-    type: String,
-    required: true
-},
-description: {
-    type: String,
-    required: true
-},
-shows:[
-    {
-        type: mongoose.Schema.Types.ObjectId,
-        ref: 'shows'
-    }
-]
+    fullName: {
+        type: String,
+        required: true
+    },
+    image: {
+        type: String,
+        required: true
+    },
+    description: {
+        type: String,
+        required: true
+    },
+    shows: [
+        {
+            type: mongoose.Schema.Types.ObjectId,
+            ref: 'Shows'
+        }
+    ]
 
 
 })
 
 
-export const ArtistModel = mongoose.models?.Artists || mongoose.model('Artists', ArtistSchema)
+const ArtistModel = mongoose.models?.Artists || mongoose.model('Artists', ArtistSchema)
+export default ArtistModel;
