@@ -4,7 +4,7 @@ export const connectToMongo = async () => {
     try {
         if (mongoose.connection.readyState === 1) {
             console.log("Already connected to MongoDB");
-            return
+            return;
         }
         await mongoose.connect(process.env.MONGO_URI)
         console.log("Connected to MongoDB");
