@@ -1,29 +1,43 @@
 const mongoose = require('mongoose')
-
+const { UserModel } = require('../users/user.model')
 
 const OrderSchema = new mongoose.Schema({
 
     userId: {
         type: mongoose.Schema.Types.ObjectId,
-        ref: 'user' //must be user
-    },
-    showName: {
-        type: String,
+        ref: 'Users', //must be user
         required: true
     },
     showId: {
         type: mongoose.Schema.Types.ObjectId,
-        ref: 'show'
+        ref: 'Shows',
+        required: true
     },
     amountOfTickets: {
         type: Number,
         required: true
     },
-    finelPrice: {
+    finalPrice: {
         type: Number,
         required: true
+    },
+    type: {
+        type: String,
+        enum: ['draft', 'payed'],
+        default: 'draft',
+        required: true,
+    },
+    purchaseDate: {
+        type: Date,
+        default: Date.now
+    },
+    userDetails: { // Optional: To store user details directly in the order
+        fullName: String,
+        phoneNumber: String,
+        email: String,
+        birthDate: String,
+        gender: String,
     }
-
 })
 
 export const OrderModel = mongoose.models?.Orders || mongoose.model('Orders', OrderSchema)
