@@ -1,24 +1,16 @@
 const mongoose = require('mongoose')
-require ('../artists/artist.model')
-require ('../users/user.model')
+ import ArtistModel from '../artists/artist.model';
+ import UserModel from '../users/user.model'; // Import if used elsewhere
 
 const ShowSchema = new mongoose.Schema({
     artist: {
+        type: mongoose.Schema.Types.ObjectId,
+        ref: 'Artists'
+    },
+    creatorName: {
         type: String,
         required: true
     },
-    // artistId: {
-    //     type: mongoose.Schema.Types.ObjectId,
-    //     ref: 'artist'
-    // },
-    creatorName: {  
-        type: String,
-        required: true
-    },
-    // creatorId: {
-    //     type: mongoose.Schema.Types.ObjectId,
-    //     ref: 'user' //must be admin
-    // },
     category: {
         type: String,
         required: true
@@ -31,12 +23,12 @@ const ShowSchema = new mongoose.Schema({
         type: String,
         required: true
     },
-    // production: {
-    //     type: String,
-    //     required: true
-    // },
-    date: {
+    production: {
         type: String,
+        required: true
+    },
+    date: {
+        type: Date,
         required: true
     },
     amount_of_tickets: {
@@ -58,4 +50,6 @@ const ShowSchema = new mongoose.Schema({
 })
 
 
-export const ShowModel = mongoose.models?.Shows || mongoose.model('Shows', ShowSchema)
+const ShowModel = mongoose.models?.Shows || mongoose.model('Shows', ShowSchema)
+export default ShowModel;
+
