@@ -1,8 +1,14 @@
-import { ShowModel } from "./show.model";
-import showController, { createShow } from "./show.controller";
+import ShowModel from "./show.model";
 
 export const getAllShows = () => {
   return ShowModel.find()
+}
+
+export const getFullShowById = async (id) => {
+  const show = await ShowModel.findById(id)
+  .populate('artist')
+  .lean(); 
+  return show;
 }
 
 export const getShowById = (id) => {
@@ -10,12 +16,13 @@ export const getShowById = (id) => {
 }
 
 export const getShowsByCategory = async (category) => {
-  return await ShowModel.find({ category: category });
+  return await ShowModel.find({ category: category })
 }
 
-export const getShowsByArtist = async (artist) =>{
-  return await ShowModel.find({artist: artist})
+export const getShowsByArtist = async (artist) => {
+  return await ShowModel.find({ artist: artist })
+  // .populate('artist')
 }
 
 
-export const createNewShowService = (show) =>  ShowModel.create(show)
+export const createNewShowService = (show) => ShowModel.create(show)
